+++ lib/src/node/error.dart
@@ -1,0 +1,11 @@
+// Copyright 2016 Google Inc. Use of this source code is governed by an
+// MIT-style license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
+import 'package:js/js.dart';
+
+@JS("Error")
+class JSError {
+  external String get message;
+  external JSError(String message);
+}